+++ tests/unit_tests/providers/async_providers/opensearch/utils_test.py
@@ -2,7 +2,7 @@ import logging
 from datetime import date
 
 from sciety_labs.providers.article_recommendation import ArticleRecommendationFilterParameters
-from sciety_labs.providers.async_providers.opensearch.opensearch_article_recommendation import (
+from sciety_labs.providers.async_providers.opensearch.utils import (
     get_article_meta_from_document,
     get_article_recommendation_from_document,
     get_from_publication_date_query_filter,
